refactor: simplify ActivityList component by removing unused functions and integrating ActivityItem component
